requirements.txt
kafka-python
pymongo[srv]
certifi
snowflake-connector-python
python-dotenv